Boas práticas de Implementação

+++ FormacaoFullStack/FundamentosAPIsASPNETCORE/ApiFuncional/ApiFuncional/Controllers/ProdutosController.cs
@@ -2,6 +2,7 @@
 using ApiFuncional.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Http;
 
 namespace ApiFuncional.Controllers
 {
@@ -17,30 +18,54 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos()
         {
             return await _context.Produtos.ToListAsync();
         }
 
         [HttpGet("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> GetProduto(int id)
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)
+            {
+                return NotFound();
+            }
 
             return produto;
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PostProduto(Produto produto)
         {
-            if(!ModelState.IsValid)
+            if (_context.Produtos == null)
+            {
+                return Problem("Erro ao criar um produto, contate o suporte!");
+            }
+
+            if (!ModelState.IsValid)
             {
                 // return BadRequest(ModelState);
 
                 // return ValidationProblem(ModelState);
 
                 return ValidationProblem(new ValidationProblemDetails(ModelState)
-                { 
+                {
                     Title = "Um ou mais erros de validação ocorreram!"
                 });
             }
@@ -54,25 +79,62 @@ namespace ApiFuncional.Controllers
         [HttpPut]
         public async Task<ActionResult> PutProduto(int id, Produto produto)
         {
-            if(id != produto.Id) return BadRequest();
+            if (id != produto.Id) return BadRequest();
 
-            if(!ModelState.IsValid) return ValidationProblem(ModelState);
+            if (!ModelState.IsValid) return ValidationProblem(ModelState);
 
-            _context.Produtos.Update(produto);
-            await _context.SaveChangesAsync();
+            // _context.Produtos.Update(produto);
+            _context.Entry(produto).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+
+                if (!ProdutoExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
 
             return NoContent();
         }
 
         [HttpDelete("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> DeleteProduto(int id)
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)
+            {
+                return NotFound();     
+            }
+
             _context.Produtos.Remove(produto);
 
             await _context.SaveChangesAsync();
 
             return NoContent();
+        }
+
+        private bool ProdutoExists(int id)
+        {
+            return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
         }
